requirements.txt
shipy-web
uvicorn[standard]
